Update Akka version
This version bump comes with two kind of modifications in code:
 - modifications so that the code compiles again after class and method
   names were changed in Akka ;
 - renaming of actor publisher to stick to the new Akka stream naming
   conventions.

This first step simply makes sure that everything compiles and works as
it used to work with Akka 2.4. There still are warnings due to
deprecation that will be addressed later.
